+++ src/threading.cpp
@@ -1,21 +1,22 @@
 #include "liblgroup.h"
+#include "threading.h"
 
-#include <Eigen/Core>
+#include <algorithm>
 #include <omp.h>
-
+#include <Eigen/Core>
 
 namespace librectify {
 
 
 #ifdef _OPENMP
-static int num_threads = omp_get_max_threads();
+static int num_threads = 0;
 #endif
 
 
 void set_num_threads(int t)
 {
     #ifdef _OPENMP
-    num_threads = t;
+    num_threads = std::min(t, omp_get_max_threads());
     Eigen::setNbThreads(t);
     #endif
 }
@@ -30,6 +31,12 @@ int get_num_threads()
     _t = 0;
     #endif
     return _t;
+}
+
+
+bool is_omp_enabled()
+{
+    return num_threads >= 0;
 }
 
 } // namespace